Refactor privacy dropdown to have an interface closer to upstream's
